feat(quill): spike single-date DatePicker with shared grid + input
Sibling DatePicker (Date in/out) alongside the range DateTimePicker, sharing
both the month grid (calendar-grid.tsx) and the segmented date/time input
(segmented-date-input.tsx) so neither is duplicated. Time-format follow-ups
(12/24h, granularity) now have a single home in the shared input.

Generated-By: PostHog Code
Task-Id: 95dcc47c-52cc-4bf9-b821-5656c41318af
